DebugProvider class created to refactor DebugServer object
