+++ metals/src/main/scala/scala/meta/internal/metals/debug/DebugServer.scala
@@ -1,36 +1,9 @@
 package scala.meta.internal.metals.debug
 
-import java.net.InetSocketAddress
-import java.net.ServerSocket
-import java.net.Socket
 import java.net.URI
-import java.{util => ju}
-import java.util.concurrent.TimeUnit
-import ch.epfl.scala.bsp4j.BuildTargetIdentifier
-import ch.epfl.scala.{bsp4j => b}
-import com.google.common.net.InetAddresses
-import com.google.gson.JsonElement
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
-import scala.concurrent.Promise
-import scala.meta.internal.metals.BuildServerConnection
-import scala.meta.internal.metals.BuildTargets
 import scala.meta.internal.metals.Cancelable
-import scala.meta.internal.metals.DefinitionProvider
-import scala.util.Failure
-import scala.util.Try
-import scala.meta.internal.metals.DebugUnresolvedMainClassParams
-import scala.meta.internal.metals.Messages.UnresolvedDebugSessionParams
-import scala.meta.internal.metals.DebugUnresolvedTestClassParams
-import scala.meta.internal.metals.BuildTargetClassesFinder
-import java.util.Collections.singletonList
-import scala.meta.internal.metals.JsonParser._
-import scala.meta.internal.metals.Compilations
-import scala.meta.internal.metals.BuildTargetClasses
-import scala.meta.internal.metals.ClassNotFoundInBuildTargetException
-import scala.meta.internal.metals.MetalsLanguageClient
-import org.eclipse.lsp4j.MessageParams
-import org.eclipse.lsp4j.MessageType
 
 final class DebugServer(
     val sessionName: String,
@@ -63,213 +36,4 @@ final class DebugServer(
     isCancelled = true
     if (proxy != null) proxy.cancel()
   }
-}
-
-object DebugServer {
-  import scala.meta.internal.metals.MetalsEnrichments._
-
-  def start(
-      parameters: b.DebugSessionParams,
-      definitionProvider: DefinitionProvider,
-      buildTargets: BuildTargets,
-      buildServer: => Option[BuildServerConnection]
-  )(implicit ec: ExecutionContext): Future[DebugServer] = {
-    Future.fromTry(parseSessionName(parameters)).flatMap { sessionName =>
-      val proxyServer = new ServerSocket(0)
-      val host = InetAddresses.toUriString(proxyServer.getInetAddress)
-      val port = proxyServer.getLocalPort
-      val uri = URI.create(s"tcp://$host:$port")
-      val connectedToServer = Promise[Unit]()
-
-      val awaitClient =
-        () => Future(proxyServer.accept()).withTimeout(10, TimeUnit.SECONDS)
-
-      // long timeout, since server might take a while to compile the project
-      val connectToServer = () => {
-        buildServer
-          .map(_.startDebugSession(parameters))
-          .getOrElse(BuildServerUnavailableError)
-          .withTimeout(60, TimeUnit.SECONDS)
-          .map { uri =>
-            val socket = connect(uri)
-            connectedToServer.trySuccess(())
-            socket
-          }
-      }
-
-      val proxyFactory = { () =>
-        val targets = parameters.getTargets.asScala
-          .map(_.getUri)
-          .map(new BuildTargetIdentifier(_))
-        val sourcePathProvider = new SourcePathProvider(
-          definitionProvider,
-          buildTargets,
-          targets.toList
-        )
-        DebugProxy
-          .open(sessionName, sourcePathProvider, awaitClient, connectToServer)
-      }
-      val server = new DebugServer(sessionName, uri, proxyFactory)
-
-      server.listen.andThen { case _ => proxyServer.close() }
-
-      connectedToServer.future.map(_ => server)
-    }
-  }
-
-  def resolveMainClassParams(
-      params: DebugUnresolvedMainClassParams,
-      buildTargetClassesFinder: BuildTargetClassesFinder,
-      compilations: Compilations,
-      buildTargets: BuildTargets,
-      buildTargetClasses: BuildTargetClasses,
-      languageClient: MetalsLanguageClient
-  )(implicit ec: ExecutionContext): Future[b.DebugSessionParams] = {
-    withRebuildRetry(
-      () =>
-        buildTargetClassesFinder
-          .findMainClassAndItsBuildTarget(
-            params.mainClass,
-            Option(params.buildTarget)
-          ),
-      compilations,
-      buildTargets,
-      buildTargetClasses
-    ).map {
-      case (clazz, target) :: others =>
-        if (others.nonEmpty) {
-          reportOtherBuildTargets(
-            clazz.getClassName(),
-            target,
-            others,
-            "main",
-            languageClient
-          )
-        }
-        clazz.setArguments(Option(params.args).getOrElse(List().asJava))
-        clazz.setJvmOptions(
-          Option(params.jvmOptions).getOrElse(List().asJava)
-        )
-        new b.DebugSessionParams(
-          singletonList(target.getId()),
-          b.DebugSessionParamsDataKind.SCALA_MAIN_CLASS,
-          clazz.toJson
-        )
-    }
-  }
-
-  def resolveTestClassParams(
-      params: DebugUnresolvedTestClassParams,
-      buildTargetClassesFinder: BuildTargetClassesFinder,
-      compilations: Compilations,
-      buildTargets: BuildTargets,
-      buildTargetClasses: BuildTargetClasses,
-      languageClient: MetalsLanguageClient
-  )(implicit ec: ExecutionContext): Future[b.DebugSessionParams] = {
-    withRebuildRetry(
-      () => {
-        buildTargetClassesFinder
-          .findTestClassAndItsBuildTarget(
-            params.testClass,
-            Option(params.buildTarget)
-          )
-      },
-      compilations,
-      buildTargets,
-      buildTargetClasses
-    ).map {
-      case (clazz, target) :: others =>
-        if (others.nonEmpty) {
-          reportOtherBuildTargets(
-            clazz,
-            target,
-            others,
-            "test",
-            languageClient
-          )
-        }
-        new b.DebugSessionParams(
-          singletonList(target.getId()),
-          b.DebugSessionParamsDataKind.SCALA_TEST_SUITES,
-          singletonList(clazz).toJson
-        )
-    }
-  }
-
-  private def parseSessionName(
-      parameters: b.DebugSessionParams
-  ): Try[String] = {
-    import scala.meta.internal.metals.JsonParser._
-    parameters.getData match {
-      case json: JsonElement =>
-        parameters.getDataKind match {
-          case "scala-main-class" =>
-            json.as[b.ScalaMainClass].map(_.getClassName)
-          case "scala-test-suites" =>
-            json.as[ju.List[String]].map(_.asScala.sorted.mkString(";"))
-        }
-      case data =>
-        val dataType = data.getClass.getSimpleName
-        Failure(new IllegalStateException(s"Data is $dataType. Expecting json"))
-    }
-  }
-
-  private def connect(uri: URI): Socket = {
-    val socket = new Socket()
-
-    val address = new InetSocketAddress(uri.getHost, uri.getPort)
-    val timeout = TimeUnit.SECONDS.toMillis(10).toInt
-    socket.connect(address, timeout)
-
-    socket
-  }
-
-  private def withRebuildRetry[A](
-      f: () => Try[A],
-      compilations: Compilations,
-      buildTargets: BuildTargets,
-      buildTargetClasses: BuildTargetClasses
-  )(implicit ec: ExecutionContext): Future[A] = {
-    Future.fromTry(f()).recoverWith {
-      case ClassNotFoundInBuildTargetException(_, buildTarget) =>
-        val target = Seq(buildTarget.getId())
-        for {
-          _ <- compilations.compileTargets(target)
-          _ <- buildTargetClasses.rebuildIndex(target)
-          result <- Future.fromTry(f())
-        } yield result
-      case _: ClassNotFoundException =>
-        val allTargets = buildTargets.all.toSeq.map(_.info.getId())
-        for {
-          _ <- compilations.compileTargets(allTargets)
-          _ <- buildTargetClasses.rebuildIndex(allTargets)
-          result <- Future.fromTry(f())
-        } yield result
-    }
-  }
-
-  private def reportOtherBuildTargets(
-      className: String,
-      buildTarget: b.BuildTarget,
-      others: List[(_, b.BuildTarget)],
-      mainOrTest: String,
-      languageClient: MetalsLanguageClient
-  ) = {
-    val otherTargets = others.map(_._2.getDisplayName())
-    languageClient.showMessage(
-      new MessageParams(
-        MessageType.Warning,
-        UnresolvedDebugSessionParams
-          .runningClassMultipleBuildTargetsMessage(
-            className,
-            buildTarget.getDisplayName(),
-            otherTargets,
-            mainOrTest
-          )
-      )
-    )
-  }
-
-  private val BuildServerUnavailableError =
-    Future.failed(new IllegalStateException("Build server unavailable"))
 }
